requirements.txt
# Core dependencies
pefile>=2023.2.7
pycryptodome>=3.19.0
requests>=2.31.0
click>=8.1.7
PyYAML>=6.0.1
colorama>=0.4.6

# Database
SQLAlchemy>=2.0.23

# Optional GUI
# PyQt6>=6.6.0
